feat: Week 9-10 cross-team integration - 26+ compilation fixes
Major achievements:
- Fix UTF-8 encoding issue (hot_exec.rs U+FFFD  U+2705)
- Complete crate::util function implementations (28 functions)
- Fix GovernorLayer generic parameters (5 E0107 errors)
- Create stub modules: src/update.rs, src/video_export.rs
- Add missing types: ToolResult, AssistantMessage, RetryConfig
- Fix 13 structural/syntax/lifetime errors across 11 files
- Resolve E0432/E0603/E0061/E0107/E0277/E0515 error categories

Stats:
- 84 files changed
- Net reduction: ~500 lines of error-prone code
- Error count reduction: ~40% (estimated)

Ready for final workspace compilation verification

+++ src/prompt.rs
@@ -35,7 +35,7 @@ impl SplitSystemPrompt {
             self.static_part.clone()
         } else {
             format!("{}\n\n{}", self.static_part, self.dynamic_part)
-        })
+        }) as usize
     }
 }
